Déplacement fonction creerOjetPizza classe Pizza -> classe PizzeriaAdminConsoleApp
