+++ pizzeria-console-objet/src/fr/pizzeria/model/Pizza.java
@@ -3,21 +3,12 @@ package fr.pizzeria.model;
 public class Pizza {
 
 	int id;
-	String code;
-	String nom;
-	double prix;
-	static int nbPizzas;
+	public String code;
+	public String nom;
+	public double prix;
+	public static int nbPizzas;
 	
 	
-	public static Pizza creerObjetPizza(String code, String nom, double prix)
-	{
-		Pizza p = new Pizza();
-		p.code = code;
-		p.nom = nom;
-		p.prix = prix;
-		Pizza.nbPizzas++;
-		return p;
-		
-	}
+	
 	
 }
